大量内容修改

+++ bootdo/src/main/java/com/bootdo/blog/dao/BCommentsMapper.java
@@ -15,13 +15,13 @@ import org.apache.ibatis.annotations.Update;
  * 
  * @author chglee
  * @email 1992lcg@163.com
- * @date 2017-09-08 17:14:06
+ * @date 2017-09-10 19:27:42
  */
 @Mapper
 public interface BCommentsMapper {
 
 	@Select("select `id`, `created`, `title`, `author`, `author_id`, `owner_id`, `email`, `url`, `ip`, `agent`, `content`, `type`, `status`, `parent` from b_comments where id = #{id}")
-	BCommentsDO get(Integer id);
+	BCommentsDO get(Long id);
 	
 	@Select("<script>" +
 	"select * from b_comments " +
